api: widen address-range helpers to 64-bit bounds; test the hardening model
kasld_addr_in_window/in_range took unsigned long bounds, which truncated
a bound wider than the platform word on 32-bit arches — MAX_PLAUSIBLE_
KERNEL_PHYS (1<<50) folded to 0, so kernel_image_phys_bound rejected
every phys witness and emitted no Q_PHYS_TEXT_BASE bound (caught by
test_engine under qemu on arm/ppc32/mips32/i686). Take the bounds as
unsigned long long; virtual-address callers' unsigned long promotes
losslessly, physical ranges no longer truncate.

Also add test_build_hardening_report: direct, deterministic coverage of
build_hardening_report() (exposure, posture, per-gate gated/blocked/
bypassed counts and names, suggestions, vuln / surface / side-channel /
no-mitigation lists), since the renderers only exercised it indirectly.
Posture is asserted per KASLR_SUPPORTED (unsupported outranks
randomization_failed on arches without KASLR).

+++ src/include/kasld/api.h
@@ -298,13 +298,20 @@ static inline unsigned long kasld_xkphys_to_phys(unsigned long va) {
  * otherwise turn the comparison into a compile-time tautology at the call site
  * (-Wlogical-op / -Wtype-limits). Routing the per-arch window checks through
  * these helpers keeps them honest and centralises the empty-window degradation.
+ *
+ * Bounds are unsigned long long: a physical-address range can legitimately
+ * exceed the platform word on 32-bit arches (e.g. MAX_PLAUSIBLE_KERNEL_PHYS =
+ * 1<<50), and an `unsigned long` parameter would silently truncate it.
+ * Virtual-address callers pass `unsigned long`, which promotes losslessly.
  * Pure arithmetic; safe in any TU. */
-static inline int kasld_addr_in_window(unsigned long addr, unsigned long lo,
-                                       unsigned long hi) {
+static inline int kasld_addr_in_window(unsigned long long addr,
+                                       unsigned long long lo,
+                                       unsigned long long hi) {
   return lo < hi && addr >= lo && addr < hi;
 }
-static inline int kasld_addr_in_range(unsigned long addr, unsigned long lo,
-                                      unsigned long hi) {
+static inline int kasld_addr_in_range(unsigned long long addr,
+                                      unsigned long long lo,
+                                      unsigned long long hi) {
   return addr >= lo && addr <= hi;
 }
